[eudsl-llvmpy] MIR: extract ScopedDiagnosticCapture/withDetail into Diagnostics.h
